requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf_yasg==1.21.7
python-decouple==3.8
